requirements.txt
# Core framework
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.10.5

# DataPizza AI components
datapizza-ai
datapizza-ai-clients-openai
datapizza-ai-vectorstores-qdrant
datapizza-ai-embedders-openai

# Vector database
qdrant-client

# Utilities
numpy>=1.24.0
pytest>=7.4.0